map: add minimal scroll cards
You can scroll through the zones, horizontally, at the bottom of the
screen now, as well as switch to a list view and remove the map. Content
of the cards is lorem ipsum basically, nothing interesting. Animations
for the map are not implemented yet, nor is some shared reference that
would connect the visible horizontal card to a highlighted map polygon.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>