api: use a client-side mapping for travel advice
We only get travel advice from the map layer and full forecast API, not
the fragment API, which makes relying on that field problematic. It
looks like the extant web application is simply using a client-side
mapping, anyway, instead of the API fields, so we can do that as well.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>