*: pre-fetch all data for the forecasts
On app load, we will now be fetching the center metadata, map layer, all
forecasts that pertain to our zones, and all media in all of the
forecasts. This will allow us to make sure that all data is present for
the user regardless of whether they've looked at all the forecasts
before going out of service.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>