*: fetch forecasts using the latest product API
Ultimately, we will need two types of views:

1. "show me the currently-valid products for a zone"
2. "show me the valid products for a zone at some point in the past"

Due to quirks of how the date filtering works on the products API from
the NAC (filtering on the publication date, not the date for which the
product is valid), we need to split these cases up. This PR does the
first pass to allow us to fetch the currently-valid forecast.
Follow-up PRs will:

1. update our routes to not contain a date, and resolve "now" when we're
   in the component
2. add a new route for a particular time in history
3. add warnings & synopses to the data we fetch for a forecast date

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>