hooks: improve prefetch logging
Previously, the logs did not show *which* data was being pre-fetched,
and that's not ideal. Also, the "the data is cached..." line was a)
superfluous with the "finished prefetching" line and b) would fire when
we issued the promise, not when it was resolved, so it was not really
useful.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>