*: use an image cache in the local filesystem for remote uris
For whatever reason, `Image.prefetch()` does not work. Also, in
developer mode our local assets are served via localhost network and in
the future some assets might come from the Expo CDN, so in general we do
not know if our assets and images come from the filesystem or from a
local file. So, we can use a cache in the app's cache directory to hold
on to these images and display only from the local filesystem.

We do not really want to have to implement a complex caching state
machine, so we basically attach ourselves to the react-query persister
by recording the local file for each URI in the react-query cache. We
can then subscribe to cache changes to clean up our local filesystem
caceh when the react-query cache has a key removed.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>