map: improve the calculations for the top elements height
Calculate the real, relative-to-window (not parent) height of the top
elements using a container that we can add the search bar to in the
future. Moved the `top={12}` into a separate view in the column since
for whatever reason the layout calculations could not handle the former,
but are fine with the latter.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>