network: periodically reconcile caches
It is possible for the cache of remote URI to filesystem location to be
out-of-sync with the real filesystem state, since we are placing our
assets into the cache directory, which the phone OS has the privilege to
clean when under filesystem pressure. Furthermore, if the app crashes or
is killed before our event-driven cleanup subscription to the
react-query cache fires, we will also be out-of-sync. Therefore, we need
to do a periodic, level-driven reconciliation to ensure we are
eventually consistent. Unfortunately there are some race conditions here
we cannot really control, but we can do "ok".

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>