map: hoist queries to the top for zones
We have a couple of different places we get data from for map zones -
the map layer query and all the individual queries for the forecasts and
warnings we do for each zone. Not only was the previous query silently
failing to do anything with the updated data (oops, setting fields that
did not exist since we were not very specific with our types), but it
also had the isseus with composed queries that we know about.

Hoisting the values up to the top allows us to not chain queries,
thereby making our error and loading states clearer, and it allows us to
more explicitly tesselate data from all of our sources.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>