observations: use a map marker to show the location
You are not supposed to be putting other types of elements into the
children of the map, and the map marker component from the library is
what we want here, anyway. We can't use the built-in `image` prop on the
marker since we need to re-size our marker; apparently there are
performance problems when there are many markers with custom image
children on one map, but we know we will only ever have one here.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>