observations: cache imprecise results
We only allow for "newest" observation fetching today, and for that we
would prefer to have stale but present cached data versus precise but
missing data. If we simply cache that we're getting *any* observations
for a day, and refresh hourly, we should make sure that users can come
back to these screens and see something useful every time, especially
since we're pre-fetching. When/if we expand to allow viewing results
from a specific range in the past, we can update those calls to use a
different key and cache, since we need to differentiate those from
"latest" calls and we can cache those precisely.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>