*: use the capabilities API to drive center selector screen
The new capabilities API lets us know which centers are registered in
the NAC system and for which cepabilities they've signed up. This PR
adds the data fetching as well as using this data to drive the center
selection screen(s). A follow-up will start to use the capabilities
enumerated to determine when we show which parts of the app.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>