weather_data: add a detail screen for one weather station
THis screen could be very similar to the one we use for multiple
stations for NWAC, but unfortunately the client-side choices for things
like:
 - what variables are named
 - how units are formatted
 - in what order data is shown

Are all unique and distinct. So, looking at one weather station's data
in the NWAC-specific view will be slightly different than in this one,
but I don't know that we can do any better for now.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>