WeatherTab: fix weather stations
The performance changes broke rendering of weather stations - as we
wanted to put the calculation of which stations needed to be shown into
a useMemo() hook, we had to hoist that up above any conditional returns.
However, we only populated the stationsByZone conditionally once we knew
we were rendering an NWAC page, so we always memoized an empty result.
This ditches the useMemo() call (but retains the removal of arrow
functions) and hopefully does not cause performance problems.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>