app.json: set runtime version policy
Our app contains dependencies outside of the Expo SDK which ship native
code. Therefore, we will be diligent to bump the runtime version of the
app whenever one of those dependencies changes. We futhermore require
that updates we ship only target the version of the app that they were
built on top of, to ensure that users on older native versions do not
receive OTA updates that are incompatible with their native layer.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>