weatherStationTimeseries: don't ask for calculated diffs on nwac
The NWAC website has besoke views of weather data, which require that
calculated differences are not requested from the Snobound endpoint. If
we ever move to make the app show weather data in line with the rest of
the NAC centers, we can remove this logic and always ask for diffs.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>