components/forecast: show toasts correctly
For whatever reason, `useTimeout()` needs to be used when calling
`Toast.show()` in a hook. Without that, we weren't seeing any of these
show up, ever. In addition, we now use the `useFocusEffect()` hook from
`react-navigation` for a much more robust approach to showing and hiding
this toast when folks navigate around the app while the toast is open.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>