PostHog: initialize a client manually
PostHog requires that the Provider be inside of / a child of the
navigation context. However, PostHog also spits out stack-traces and
yells when the client is re-initialized, so having their Provider in
part of the App that gets re-rendered only leads to spam. Initialize the
client manually at the top-level so we don't see any spam.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>