map: only overwrite danger when we get a newer forecast result
When we ask the NAC for the absolutely latest forecast product for a
zone, it is possible for one zone to have a product that is valid for
many zones. If that product is older than other, more zone-specific
products, this will poison our `forecastResults` and the order in which
we process this data will matter. Never use stale results when we have
more recent ones already from the map layer.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>