AvalancheForecastZoneMap: update watch/bulletin behavior
The warnings endpoint can return three different types of products
(warnings, watches and special bulletins). We only want the map to flash
when there's an active warning.

A small fix to add 1 day to the warning fetch hook for the time machine,
since the NAC API assumes a date and appends a time to it.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>