*: remove observations feed date limits
The observations feed doesn't have any good reason to enforce a bound on
how far back in time you can scroll. This paradigm was ported from the
website where choosing "last season" will shoot off a request that
fetches all observations for the entire season. Since we're not asking
for all observations at once and doing a progressive load, there's no
real reason to have this behavior in the app.

Signed-off-by: Steve Kuznetsov <stekuznetsov@microsoft.com>