observations: fix form values (#846)
It looks like bumping the react-hook-form library changed the semantics
of data in default values for a form that was never bound to a
controller. In the past, this data remained in the values of the form,
but now it does not. We were failing to validate the observation state
before submitting it as we were losing the `status` field. Binding
invisible and unchanging values to controllers that are disabled and not
shown to the user is a simple if inelegant workaround.

Signed-off-by: Steve Kuznetsov <stekuznetsov@microsoft.com>