*: misc updates (#868)
* patches: update react-native-render-html

Upstream seems unable to publish a new version.

Signed-off-by: Steve Kuznetsov <stekuznetsov@microsoft.com>

* components/screens: remove import cycle

The AvalancheForecast component is now only a screen, and we had an
import cycle between it and the forecast screens. This commit moves all
the screen-related components to the screens directory to break the
cycle.

Signed-off-by: Steve Kuznetsov <stekuznetsov@microsoft.com>

* patches: update tab view

Fixes the 'Warning: A props object containing a "key" prop is being
spread into JSX' issue.

Signed-off-by: Steve Kuznetsov <stekuznetsov@microsoft.com>

* package.json: bump required dependencies

Signed-off-by: Steve Kuznetsov <stekuznetsov@microsoft.com>

* components: fix typo in comment

Signed-off-by: Steve Kuznetsov <stekuznetsov@microsoft.com>

---------

Signed-off-by: Steve Kuznetsov <stekuznetsov@microsoft.com>