fix: Fixes scrolling on cards
Fixes the scrolling behaviour of the `AnimatedCards` component on
Android. The `scrollEventThrottle` parameter that had previously been
added was causing an issue with the cards scrolling on Android devices.
This fixes the issue by removing the continuous calls to `handleScroll`,
in favour of calling it a single time when the `onMomentumScrollEnd`
event fires.