Fixes scrolling on cards (#910)
* fix: Fixes typos

* fix: Fixes scrolling on cards

Fixes the scrolling behaviour of the `AnimatedCards` component on
Android. The `scrollEventThrottle` parameter that had previously been
added was causing an issue with the cards scrolling on Android devices.
This fixes the issue by removing the continuous calls to `handleScroll`,
in favour of calling it a single time when the `onMomentumScrollEnd`
event fires.

* fix: Fixes scrolling on cards

Updates scrolling on the `AnimatedCards` component so that it
continuously updates while a user is scrolling. This fixes a lag in UI
updates to the map making it feel more responsive.

* fix: Adds missing dependency to useCallback