fix: observations list view
Fixes handling of case when there is no items in the list. Should
display a loading spinner if the app is currently fetching data,
otherwise displays a warning about filtering out all data. Fixes issue #777.

Fixes display of list footer. Should display a loading spinner if the
app is currently fetching more data, otherwise the number of
observations in the list. Fixes #871.