Fixes observations list view (#942)
* fix: observations list view

Fixes handling of case when there is no items in the list. Should
display a loading spinner if the app is currently fetching data,
otherwise displays a warning about filtering out all data. Fixes issue #777.

Fixes display of list footer. Should display a loading spinner if the
app is currently fetching more data, otherwise the number of
observations in the list. Fixes #871.

* fix: Observations list view

Updates `fetchMoreData` function to take filters into account when
counting observations and adds a date limit of 6 months in the past.
When a user adds filters it should keep fetching data until it finds
something to show the user or hits the date limit.

* fix: Observations list view

Updates default date limit to one year for filters with no results.

Adds a button to increase date range when filter has no results.