*: load fonts using config plugin, use platform-specific names
We are experiencing difficulties with rendering text on Android,
specifically for non-standard fonts (italics, bolds). As a
shot-in-the-dark approach, we also noticed that we are loading fonts at
runtime for no real reason, so we are trying to load all the fonts
upfront using a config plugin instead, hoping it might help our
rendering issue. We need to handle the difference in Android and iOS
naming conventions for the fonts - Android uses the file name, iOS the
PostScript name.

Signed-off-by: Steve Kuznetsov <stekuznetsov@microsoft.com>